Added score_screen() and save_score()
Made the highscore screen and functions to save the score when game-over. It checks the players current score agains each entry of the highscore list, and if it finds that the current is higher, it moves all the entries below one spot down. Added a score int in the highscore struct, its easier to compare two integers instead of two strings/char array.